requirements.txt
bs4
black
pyrogram
playwright
playwright-stealth
python-dotenv
requests
shortuuid
tgcrypto
uvloop
urltitle